Deprecate TR_S390MachineType in favor of an integer machine id
Avoid using an enumerated `TR_S390MachineType` type to represent
machine ids to future proof the JVM running on a more modern processor
for which we do not know the machine id for.

A typical example of this is running an older JVM, say very early Java
8 snapshots of OpenJ9 which did not have z14 support. In such cases the
machine id for z14 would not match any of the existing machine ids in
the codebase at the time and we would start reporting invalid machine
ids in the verbose log ("0" to be precise).

In this change we use the true machine id as parsed from z/OS or Linux
on Z and display the number that is returned as an integer. The actual
machine detection logic relies on the port library which uses STFLE
bits to determine the processor we are executing on. This means
processor detection is unaffected by our changes here.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>