Remove hardcoded system property java.util.prefs.PreferencesFactory
This only affects J9VM_UNIX platforms.
This modification conforms to specification, and allows a
PreferencesFactory implementation from a jar file containing a
provider-configuration file named java.util.prefs.PreferencesFactory in
the resource directory META-INF/services when this system property is
not set.
In case neither the system property nor an extension jar file is
provided, a default PreferencesFactory implementation for the underlying
platform is provided. For a platform not Windows or OS X, a default
PreferencesFactory is java.util.prefs.FileSystemPreferencesFactory which
was hardcoded via system property java.util.prefs.PreferencesFactory.


Signed-off-by: Jason Feng <fengj@ca.ibm.com>