Anchor arguments of MethodHandle invoke before expansion
The expansion will insert calls before the original call. Anchor the
arguments such that the evaluation of arguments is before the execution
of the call.

Notice that missing the anchoring does not cause functional issues
because the inserted calls are internal to the JIT and have no side
effect on arguments. However, it prevents us from folding argument that
is a static final field because the inserted calls are unsupported
potential OSR points.

Signed-off-by: Liqun Liu <liqunl@ca.ibm.com>