Fold use64BitRegsOn32Bit as true on 32-bit Z targets
After the removal of `TR_GPR64` disabling the use of 64-bit registers
on 31-bit is no longer functionally correct due to the very complex and
surgical removal that took place. As the use of 64-bit registers on
31-bit had been enabled by default for several months now this marks
the perfect opportunity to fold away the code related to the query on Z
as leaving it in place in a non-functional state is not a good idea.

Along with the folding of this query on Z (which is now always true on
32-bit and always false on 64-bit for Z) we fold away all the resulting
code and evaluators which are no longer used, thus improving the
overall codebase by making it much easier to understand.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>